+++ src/shared/log-template.js
@@ -1,18 +1,28 @@
 const HTML_SUFFIX = "</div></div></body></html>";
+const LEGACY_LOG_HEAD = [
+  "<link rel=\"stylesheet\" href=\"https://fonts.googleapis.com/css?family=Source+Code+Pro|Quantico:400,400italic,700|Roboto+Mono|Comic+Mono\">",
+  "<base href=\"https://play.sindome.org\">",
+  "<link rel=\"stylesheet\" type=\"text/css\" href=\"https://www.sindome.org/css/dome.css\">",
+  "<link rel=\"stylesheet\" type=\"text/css\" href=\"https://play.sindome.org/css/client.css\">",
+].join("");
 
 /**
  * Wraps buffer markup in the HTML shell used for downloadable logs.
  *
  * @param {string | undefined | null} bufferHtml
  * @param {string | undefined | null} styleCss
+ * @param {boolean | undefined | null} inlineCss
  * @returns {string}
  */
-export function buildLogHtml(bufferHtml = "", styleCss = "") {
+export function buildLogHtml(bufferHtml = "", styleCss = "", inlineCss = true) {
   const safeBuffer = bufferHtml ?? "";
   const safeStyle = (styleCss ?? "").replace(/<\/style/gi, "<\\/style");
+  const styleMarkup = inlineCss
+    ? `<style>${safeStyle}</style>`
+    : LEGACY_LOG_HEAD;
   const htmlPrefix = [
     "<html><head><meta charset=\"utf-8\"><title>Web Client Buffer</title>",
-    `<style>${safeStyle}</style>`,
+    styleMarkup,
     "</head><body><div id=\"browser-client\"><div id=\"lineBuffer\">"
   ].join("");
   return htmlPrefix + safeBuffer + HTML_SUFFIX;